requirements.txt
# Core dependencies
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.2
cryptography>=40.0.0
plotly>=5.13.0
scikit-learn>=1.3.0

# Neural Network & Transformers
transformers>=4.31.0
sentence-transformers>=2.2.2
timm>=0.9.2

# Imaging & Video
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
albumentations>=1.3.1

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1

# Data Processing & Analysis
pandas>=2.0.3
scipy>=1.11.0
networkx>=3.1.0
h5py>=3.9.0

# Visualization
seaborn>=0.12.2

# Utilities
tqdm>=4.65.0
pyyaml>=6.0.1

# Testing & Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.1
isort>=5.12.0

# Security
pyjwt>=2.6.0
python-jose>=3.3.0

# Documentation
sphinx>=6.1.0
mkdocs>=1.4.0

# Optional Dependencies (Uncomment if needed)
# tensorboard>=2.13.0  # For experiment tracking
# wandb>=0.15.5       # For experiment monitoring
# ray>=2.6.1         # For distributed computing
# pytorch-lightning>=2.0.6  # For structured training

# Version Constraints
setuptools>=68.0.0
wheel>=0.40.0
pip>=23.2.1
